refactor(agent): 壳里的纯显示换算搬进 residentShellDisplay，浮层几何搬出组件文件
合并 main 之后撞上两道门岗，都是"两边各自长了几行"叠出来的，修在该修的层上：

- check:filesize：ProjectAgentResidentShell.tsx 806 行越过 800。把面名 / 状态名 /
  介入槽行 / 出错文案这几件**与 React 无关**的换算搬进 resident/residentShellDisplay.ts。
  residentItemClassName 刻意留在壳里：tailwind.config.ts 的 content 只扫 src 下的 .tsx，
  类名字符串搬进 .ts 就不再被生成，而且是静默的——实测搬过去之后用户气泡的
  ml-auto / max-w-[86%] 当场消失，气泡从右侧小卡变成整行通栏，12 张视觉基线同时翻红。
  两个文件顶部都写清了这条，config 本身的洞另开一单收。
- lint:ci 棘轮 82：AnchoredPopover.tsx 同时导出组件与纯函数，触发 react-refresh。
  几何本来就该是不含 React 的一份（它的测试早就单独一文件），抽成
  design/anchoredPopoverPlacement.ts，警告在根上消失而不是被豁免。
