+++ src/design/index.ts
@@ -44,8 +44,10 @@ export type { NomiIdentityIconSource } from './NomiIdentityIcon'
 export { NomiSegmented } from './NomiSegmented'
 export type { NomiSegmentedProps, NomiSegmentedOption } from './NomiSegmented'
 export { BodyPortal } from './portal'
-export { AnchoredPopover, resolveAnchoredPopoverPlacement } from './AnchoredPopover'
-export type { AnchoredPopoverProps, AnchoredPopoverAlign } from './AnchoredPopover'
+export { AnchoredPopover } from './AnchoredPopover'
+export type { AnchoredPopoverProps } from './AnchoredPopover'
+export { resolveAnchoredPopoverPlacement } from './anchoredPopoverPlacement'
+export type { AnchoredPopoverAlign } from './anchoredPopoverPlacement'
 export { hasOpenDialogAbove, NOMI_OVERLAY_Z_INDEX } from './overlayLayers'
 export { nomiDesignTokens } from './tokens'
 export { buildNomiTheme } from './theme'
