+++ src/workbench/ai/ProjectAgentResidentShell.tsx
@@ -1,4 +1,3 @@
-import type { TranslationKey } from '../../i18n/translationKey'
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 import {
@@ -11,7 +10,7 @@ import {
 import { NomiLogoMark, NomiSegmented, StatusBadge, WorkbenchIconButton } from '../../design'
 import { cn } from '../../utils/cn'
 import type { AgentToolProfile } from '../../../electron/shared/projectAgentContracts'
-import { useWorkbenchStore, type ProjectAgentReference, type ProjectAgentRunMode, type WorkspaceMode } from '../workbenchStore'
+import { useWorkbenchStore, type ProjectAgentReference, type ProjectAgentRunMode } from '../workbenchStore'
 import { useGenerationCanvasStore } from '../generationCanvas/store/generationCanvasStore'
 import { runWorkbenchAgent, type ToolCallEvent } from './workbenchAgentRunner'
 import { stopProjectAgentTurn } from './projectAgentTurnCommands'
@@ -42,7 +41,7 @@ import { ResidentReferenceChip } from './resident/ResidentReferenceChip'
 import { MenuCopy, MenuRow, Popover, PROMPT_PRESETS, ResidentPromptMenu, iconControlClass } from './resident/ResidentMenus'
 import { attachmentPayloads, itemRef } from './resident/agentItemHelpers'
 import { normalizeResidentToolProjection, readResidentToolProjections, residentToolProjectionKey, residentToolProjectionScope, writeResidentToolProjections, type ResidentToolProjection } from './resident/residentToolProjection'
-import { proposalForTool, readableToolDetailRows, readableToolName, readableToolPreview, readableToolResult, readableToolSummary, readableToolTarget, residentToolProjectionForCall } from './resident/residentToolDisplay'
+import { proposalForTool, readableToolName, readableToolPreview, readableToolResult, readableToolSummary, readableToolTarget, residentToolProjectionForCall } from './resident/residentToolDisplay'
 import { isGenerationProposalTool, proposalDecisionPayload } from './resident/generationProposalEditing'
 import { residentQuestionOptions } from './resident/residentExceptionProjections'
 import { useAssetPool } from '../assets/useAssetPool'
@@ -66,10 +65,22 @@ import { GenerationProposalEditor } from './resident/GenerationProposalEditor'
 import ReconcileDeviationCard from '../generationCanvas/components/ReconcileDeviationCard'
 import { buildContentFixMessage, useShotVerifyStore } from '../generationCanvas/agent/shotVerifyStore'
 import { isEmptyStoryboardPlan } from '../generationCanvas/agent/storyboardPlan'
+import { friendlyError, interventionDetails, isActiveQueueStatus, statusLabel, surfaceLabel, type ResidentSurface } from './resident/residentShellDisplay'
 
-type ResidentSurface = Extract<WorkspaceMode, 'creation' | 'storyboard' | 'generation' | 'preview'>
 const isDocumentSurface = (surface: ResidentSurface): boolean => surface === 'creation' || surface === 'storyboard'
 type PendingTool = { call: ToolCallEvent; bindingKey: string; state: ResidentApprovalState }
+
+/**
+ * 每一条消息挂什么样式。它**必须**留在这个 .tsx 里：tailwind.config.ts 的 content 只扫
+ * src 目录下的 .tsx，类名字符串搬进 .ts 就不再被生成，且是静默的（见 residentShellDisplay.ts 顶部）。
+ */
+function residentItemClassName(item: ProjectAgentItem, declined: boolean): string {
+  if (item.kind === 'user') return 'ml-auto min-h-[52px] max-w-[86%] text-caption text-nomi-paper'
+  if (item.kind === 'assistant') return 'max-w-full px-1 text-caption leading-5'
+  const ownsCard = (item.kind === 'failure' && !declined) || (item.kind === 'artifact' && (item.status === 'running' || item.status === 'failed'))
+  if (ownsCard) return 'max-w-full'
+  return cn('rounded-nomi-sm border px-2.5 py-1.5 text-caption', declined ? 'border-nomi-line-soft bg-nomi-ink-05' : 'border-nomi-line-soft bg-nomi-paper')
+}
 type MenuId = 'attachments' | 'references' | 'skills' | 'prompts' | 'modes' | 'policy' | 'models' | null
 
 const residentPendingTools = new Map<string, PendingTool>()
@@ -105,53 +116,6 @@ function useResidentPendingTools(key: string | null): PendingTool[] {
   const [, redraw] = React.useState(0)
   React.useEffect(() => { const listener = () => redraw((value) => value + 1); residentPendingListeners.add(listener); return () => { residentPendingListeners.delete(listener) } }, [])
   return key ? Array.from(residentPendingTools.values()).filter((item) => item.bindingKey === key) : []
-}
-
-function surfaceLabel(t: (key: string, options?: Record<string, unknown>) => string, surface: ResidentSurface): string {
-  return surface === 'generation'
-    ? t('agentResident.contextGeneration')
-    : surface === 'preview'
-      ? t('agentResident.contextPreview')
-      : surface === 'storyboard'
-        ? t('agentResident.contextStoryboard')
-        : t('agentResident.contextCreation')
-}
-
-const STATUS_LABEL_KEY = {
-  drafting: 'agentResident.planning',
-  proposed: 'agentResident.waitingApprovalShort',
-  declined: 'agentResident.declined',
-  queued: 'agentResident.queued',
-  running: 'agentResident.running',
-  done: 'agentResident.done',
-  failed: 'agentResident.failed',
-  stopped: 'agentResident.stopped',
-} as const satisfies Record<ProjectAgentStatus, TranslationKey>
-
-function statusLabel(t: (key: string, options?: Record<string, unknown>) => string, status: ProjectAgentStatus): string {
-  return t(STATUS_LABEL_KEY[status])
-}
-
-function isActiveQueueStatus(status: ProjectAgentStatus): boolean {
-  return status === 'queued' || status === 'proposed' || status === 'running'
-}
-
-function interventionDetails(t: (key: string, options?: Record<string, unknown>) => string, call: ToolCallEvent, args: Record<string, unknown>, proposal?: ReturnType<typeof proposalForTool>): readonly { label: string; value: string }[] {
-  const rows = [...readableToolDetailRows(t, call.toolName, args), ...(proposal?.fields ?? [])]
-  return rows.filter((row, index, all) => all.findIndex((candidate) => candidate.label === row.label && candidate.value === row.value) === index).slice(0, 12).map((row) => ({ label: row.label, value: row.value }))
-}
-
-function friendlyError(error: unknown, t: (key: string, options?: Record<string, unknown>) => string): string {
-  const code = error instanceof Error ? error.message : ''
-  return code === 'project_agent_unavailable' || code === 'project_binding_stale' ? t('agentResident.unavailable') : t('agentResident.sendFailed')
-}
-
-function residentItemClassName(item: ProjectAgentItem, declined: boolean): string {
-  if (item.kind === 'user') return 'ml-auto min-h-[52px] max-w-[86%] text-caption text-nomi-paper'
-  if (item.kind === 'assistant') return 'max-w-full px-1 text-caption leading-5'
-  const ownsCard = (item.kind === 'failure' && !declined) || (item.kind === 'artifact' && (item.status === 'running' || item.status === 'failed'))
-  if (ownsCard) return 'max-w-full'
-  return cn('rounded-nomi-sm border px-2.5 py-1.5 text-caption', declined ? 'border-nomi-line-soft bg-nomi-ink-05' : 'border-nomi-line-soft bg-nomi-paper')
 }
 
 type ResidentSendContext = Readonly<{
